+++ Exam_Helper/Exam_Helper.csproj
@@ -5,6 +5,17 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Views\UserLibrary\Packs\**" />
+    <Compile Remove="Views\UserLibrary\Questions\**" />
+    <Content Remove="Views\UserLibrary\Packs\**" />
+    <Content Remove="Views\UserLibrary\Questions\**" />
+    <EmbeddedResource Remove="Views\UserLibrary\Packs\**" />
+    <EmbeddedResource Remove="Views\UserLibrary\Questions\**" />
+    <None Remove="Views\UserLibrary\Packs\**" />
+    <None Remove="Views\UserLibrary\Questions\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.9" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="3.1.9" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.9" />
@@ -16,11 +27,6 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.4" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.4" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL.Design" Version="1.1.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Views\PublicLibrary\Packs\" />
-    <Folder Include="Views\PublicLibrary\Questions\" />
   </ItemGroup>
 
 </Project>
