+++ Exam_Helper/Views/UserLibrary/QCreate.cshtml
@@ -17,14 +17,14 @@
 <form asp-action="QCreate">
     <div asp-validation-summary="ModelOnly" class="text-danger"></div>
     <div class="d-flex justify-content-between pt-3 pb-3">
-        <span><b>Наименование вопроса<span class="colortext" style="color:red">*</span>:</b><input asp-for="@Model.question.Definition" type="text" class="form-control w-100"></span>
+        <span><b>Наименование вопроса<span class="colortext" style="color:red">*</span>:</b><input asp-for="@Model.question.Title" type="text" class="form-control w-100"></span>
     </div>
 
     <div class="d-flex justify-content-between pt-3 pb-3">
         <span>
             <b>Формулировка вопроса<span class="colortext" style="color:red">*</span>:</b>
             <br />
-            <textarea asp-for="@Model.question.Title" id="textarea" style="width:70vh;margin-top:2vh;" rows="3"></textarea>
+            <textarea asp-for="@Model.question.Definition" id="textarea" style="width:70vh;margin-top:2vh;" rows="3"></textarea>
         </span>
     </div>
