+++ Exam_Helper/ViewsModel/Libs/ClassForQuestionCreatingModel.cs
@@ -9,6 +9,8 @@ namespace Exam_Helper.ViewsModel
     {
         public Question question { get; set; }
         public List<TagForQuestionCreatingModel> tags { get; set; }
+
+        
     }
 
     public class TagForQuestionCreatingModel
